Refactor SQL Server data source handling for partitioned tables
